requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
greenlet==3.0.1
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9
pgvector==0.2.4
redis==5.0.1
hiredis==2.2.3
python-telegram-bot==20.7
python-multipart==0.0.6
openai>=1.40.0,<2.0.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gspread==5.12.0
aiohttp==3.9.1
duckduckgo-search==3.9.6
requests==2.31.0
email-validator==2.1.0
tenacity==8.2.3
pytz==2023.3
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
cryptography==41.0.7
httpx==0.25.2
jsonschema==4.20.0
python-dateutil==2.8.2
structlog==23.2.0
async-timeout==4.0.3